README description added, packages naming fixed
