+++ README.md
@@ -1,2 +1,73 @@
 # Sprint_4
+# Проект: Автоматизация тестирования заказа самоката
 
+Проект представляет собой автоматизированные тесты для веб-приложения,для оформления заказа на самоката. Тесты написаны на Java с использованием Selenium и JUnit.
+
+## Описание проекта
+
+Проект включает в себя автоматизированные тесты для проверки функциональности веб-приложения:
+- Заполнение данных пользователя.
+- Выбор станции метро.
+- Выбор даты и времени аренды.
+- Выбор цвета самоката.
+- Оформление заказа и проверка подтверждения.
+
+Тесты написаны с использованием паттерна Page Object Model (POM), что делает код более структурированным и поддерживаемым.
+
+## Технологии
+
+- **Java** — язык программирования.
+- **Selenium WebDriver** — для автоматизации взаимодействия с веб-браузером.
+- **JUnit** — для написания и запуска тестов.
+- **Maven** — для управления зависимостями и сборки проекта.
+- **ChromeDriver** / **GeckoDriver** — для управления браузерами Chrome и Firefox.
+
+### Основные классы
+
+- **`HomePage`** — класс для работы с главной страницей.
+- **`PersonDetaislOrderPage`** — класс для работы со страницей ввода личных данных.
+- **`AboutRentOrderPage`** — класс для работы со страницей оформления аренды.
+
+### Тесты
+
+- **`HomePageLoadTest`** — проверка загрузки главной страницы.
+- **`HeaderOrderButtonTest`** — проверка оформления заказа через кнопку "Заказать" в хедере страницы.
+- **`MiddleOrderButtonTest`** — проверка оформления заказа через кнопку "Заказать" в середине страницы.
+
+## Зависимости
+
+Проект использует следующие зависимости:
+
+- Selenium Java 3.141.59
+- JUnit 4.13.2
+- maven 2.22.2
+- WebDriverManager 5.4.1
+
+
+## Как запустить тесты
+
+1. Убедитесь, что у вас установлены:
+   - Java (JDK 11 или выше)
+   - Maven
+   - Браузер Chrome или Firefox
+
+2. Настройка окружения
+
+- Chrome
+
+Убедитесь, что у вас установлен Chrome.
+Установите chromedriver той же версии что и Chrome
+
+- Firefox
+
+Убедитесь, что у вас установлен Firefox.
+Установите geckodriver той же версии что и Firefox
+
+3. Запуск тестов
+
+- Запуск теста для Chrome
+./mvn test -Dbrowser=chrome
+
+
+- Запуск теста для Firefox
+./mvn test -Dbrowser=firefox
