+++ src/test/java/ru/header/oder/button/test/HeaderOderButtonTest.java
@@ -1,8 +1,8 @@
-package MiddleOrderButtonTest;
+package ru.header.oder.button.test;
 
-import Pages.AboutRentOrderPage;
-import Pages.HomePage;
-import Pages.PersonDetaislOrderPage;
+import ru.pages.AboutRentOrderPage;
+import ru.pages.HomePage;
+import ru.pages.PersonDetaislOrderPage;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -10,13 +10,10 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxOptions;
-import org.openqa.selenium.firefox.FirefoxProfile;
 
 
 @RunWith(Parameterized.class)
-public class MiddleOrderButtonTest {
+public class HeaderOderButtonTest {
     private WebDriver driver;
 
     // Параметры
@@ -31,7 +28,7 @@ public class MiddleOrderButtonTest {
     private String commentInput;
 
     // Конструктор для параметров
-    public MiddleOrderButtonTest(String nameInput, String surnameInput, String adressInput, String metroStation,
+    public HeaderOderButtonTest(String nameInput, String surnameInput, String adressInput, String metroStation,
                                  String phoneInput, String pickDate, String pickTime, String blackColorCheckBox, String commentInput) {
         this.nameInput = nameInput;
         this.surnameInput = surnameInput;
@@ -54,13 +51,14 @@ public class MiddleOrderButtonTest {
 
     @Before
     public void setUp() {
-//              CHROME DRIVER
+
+ //       CHROME DRIVER
         String driverPath = "chromedriver";
         System.setProperty("webdriver.chrome.driver", driverPath);
         driver = new ChromeDriver();
     }
 
-////              FIREFOX DRIVER
+////          FIREFOX DRIVER
 //        // Укажите путь к geckodriver, если он не в PATH
 //        System.setProperty("webdriver.gecko.driver", "geckodriver");
 //
@@ -74,20 +72,24 @@ public class MiddleOrderButtonTest {
 //        // Инициализация FirefoxDriver
 //        driver = new FirefoxDriver(options);
 //
-//}
+//    }
 
     @Test
-    public void FillOrderDetailsPageMiddle() {
+    public void FillOrderDetailsPageHeader() {
         HomePage objHomePage = new HomePage(driver);
         objHomePage.isHomePageLoaded();
         PersonDetaislOrderPage objPersonDetaislOrderPage = new PersonDetaislOrderPage(driver);
 
-        objHomePage.scrollToMiddleOrderButton();
-        objHomePage.clickOrderButtonMiddle();
+        objHomePage.clickOrderButtonHeader();
         objPersonDetaislOrderPage.isPersonDetaislPageLoaded();
 
+        // переменная со значением, для поля ИМЯ
+//        String nameInput = "Иван";
+        // в одном шаге проверь, что поле «Занятие» доступно для редактирования, и введи в него новое значение
         objPersonDetaislOrderPage.emptyNameInput(nameInput);
+//        String surnameInput = "Крузенштерн";
         objPersonDetaislOrderPage.emptySurnameInput(surnameInput);
+//        String adressInput = "Семеновская роща 15";
         objPersonDetaislOrderPage.emptyAdressInput(adressInput);
 
 //        objPersonDetaislOrderPage.clickMetroSelector();
@@ -102,8 +104,12 @@ public class MiddleOrderButtonTest {
         }
 
 
+        // проверь, что поле «Станция метро» поменяло значение на новое
+//        String newMetroStation = "Красносельская";
+ //       objPersonDetaislOrderPage.waitForChangMetroStation(newMetroStation);
 
-//        objPersonDetaislOrderPage.waitForChangMetroStation(metroStation);
+        // проверь, что поле «Телефон»  поменяло значение на новое
+ //       String phoneInput = "+79213458909";
         objPersonDetaislOrderPage.emptyPhoneInput(phoneInput);
 
         objPersonDetaislOrderPage.clickNextButton();
@@ -111,6 +117,7 @@ public class MiddleOrderButtonTest {
         AboutRentOrderPage objAboutRentOrderPage = new AboutRentOrderPage(driver);
         objAboutRentOrderPage.isAboutRentOrderPageLoad();
 
+        // Проверка текста заголовка
         String expectedHeaderText = "Про аренду";
         String actualHeaderText = objAboutRentOrderPage.getHeaderText();
         if (actualHeaderText.equals(expectedHeaderText)) {
@@ -119,12 +126,20 @@ public class MiddleOrderButtonTest {
             System.out.println("Текст заголовка некорректен. Ожидалось: " + expectedHeaderText + ", но найдено: " + actualHeaderText);
         }
 
+
+//        objAboutRentOrderPage.clickRentDateInput();
+//        objAboutRentOrderPage.clickPickDate();
         objAboutRentOrderPage.clickRentDateInput();
         if (pickDate.equals("17")) {
             objAboutRentOrderPage.clickPickDate();
         } else if (pickDate.equals("18")) {
             objAboutRentOrderPage.clickPickDate2();
         }
+
+//        objAboutRentOrderPage.clickRentTimeInput();
+//        //       String dateInput = "13.03.2025";
+//        //       objAboutRentOrderPage.emptyRentDateInput(dateInput);
+//        objAboutRentOrderPage.clickPickTime();
 
         objAboutRentOrderPage.clickRentTimeInput();
         if (pickTime.equals("сутки")) {
@@ -133,6 +148,7 @@ public class MiddleOrderButtonTest {
             objAboutRentOrderPage.clickPickTime2();
         }
 
+//        objAboutRentOrderPage.clickBlackColorCheckBox();
         if (blackColorCheckBox.equals("чёрный жемчуг")) {
             objAboutRentOrderPage.clickBlackColorCheckBox();
         } else if (blackColorCheckBox.equals("серая безысходность")) {
@@ -151,12 +167,14 @@ public class MiddleOrderButtonTest {
             System.out.println("серая безысходность' не выбран.");
         }
 
+ //       String commentInput = "Подъезд со двора";
         objAboutRentOrderPage.emptyCommentInput(commentInput);
 
         objAboutRentOrderPage.clickOrderButton();
 
         objAboutRentOrderPage.isOrderConfirmationHeader();
         objAboutRentOrderPage.clickConfirmOrderButton();
+
 
         objAboutRentOrderPage.getConfirmationMessageHeaderText();
 
@@ -169,18 +187,11 @@ public class MiddleOrderButtonTest {
             System.out.println("Текст заголовка некорректен. Ожидалось: " + expectedConfirmationMessageHeaderText + ", но найдено: " + actualConfirmationMessageHeaderText);
         }
 
-        objAboutRentOrderPage.checkStatusButtonText();
-        String expectedButtonText = "Посмотреть статус";
-        String actualButtonText = objAboutRentOrderPage.checkStatusButtonText();
-        if (actualButtonText.equals(expectedButtonText)) {
-            System.out.println("Текст кнопки корректен: " + actualButtonText);
-        } else {
-            System.out.println("Текст кнопки некорректен. Ожидалось: " + expectedButtonText + ", но найдено: " + actualButtonText);
-        }
-
         objAboutRentOrderPage.clickCheckStatusButton();
 
     }
+
+
 
     @After
     public void tearDown() {
